feat(strm-812): add strm-cli-version header to every request
