+++ pkg/cmd/version.go
@@ -3,19 +3,16 @@ package cmd
 import (
 	"fmt"
 	"github.com/spf13/cobra"
+	"strmprivacy/strm/pkg/common"
 )
-
-var GitSha = "dev"
-var Version = "dev"
-var BuiltOn = "unknown"
 
 var VersionCmd = &cobra.Command{
 	Use:               "version",
 	DisableAutoGenTag: true,
 	Short:             "Print CLI version",
 	Run: func(cmd *cobra.Command, args []string) {
-		println(fmt.Sprintf("STRM Privacy CLI version: %s", Version))
-		println(fmt.Sprintf("Git commit: %s", GitSha))
-		println(fmt.Sprintf("Built on: %s", BuiltOn))
+		println(fmt.Sprintf("STRM Privacy CLI version: %s", common.Version))
+		println(fmt.Sprintf("Git commit: %s", common.GitSha))
+		println(fmt.Sprintf("Built on: %s", common.BuiltOn))
 	},
 }
